Sawmill - Re-add some occluders
Re-added some occluders from the stock map that i removed because i thought it was bad for some reason
Also added a few more occluders of my own and improved the stock ones.

Also fixed a few solids that hammer++ now reports as invalid, only on koth because i'm lazy.